+ Added selection of stick camera.

+++ llc/gpk_coord.h
@@ -10,8 +10,8 @@ namespace gpk
 {
 #pragma pack(push, 1)	// You can read about pragma pack() here: https://www.google.com/search?q=pragma+pack
 	struct SNearFar {
-												double			Near		;
-												double			Far			;
+												float			Near		;
+												float			Far			;
 	};
 
 	template<typename _tBase>
